make recycler static to avoid OOM problem
