+++ bookkeeper-server/src/main/java/org/apache/bookkeeper/proto/CompletionKey.java
@@ -1,0 +1,13 @@
+package org.apache.bookkeeper.proto;
+
+import org.apache.bookkeeper.proto.BookkeeperProtocol.OperationType;
+
+abstract class CompletionKey {
+    OperationType operationType;
+
+    CompletionKey(OperationType operationType) {
+        this.operationType = operationType;
+    }
+
+    public void release() {}
+}
