+++ bookkeeper-server/src/main/java/org/apache/bookkeeper/proto/PerChannelBookieClient.java
@@ -165,7 +165,7 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
     static final Logger LOG = LoggerFactory.getLogger(PerChannelBookieClient.class);
 
     // this set contains the bookie error return codes that we do not consider for a bookie to be "faulty"
-    private static final Set<Integer> expectedBkOperationErrors = Collections.unmodifiableSet(Sets
+    protected static final Set<Integer> expectedBkOperationErrors = Collections.unmodifiableSet(Sets
             .newHashSet(BKException.Code.BookieHandleNotAvailableException,
                         BKException.Code.NoSuchEntryException,
                         BKException.Code.NoSuchLedgerExistsException,
@@ -201,79 +201,79 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
         name = BookKeeperClientStats.CHANNEL_READ_OP,
         help = "channel stats of read entries requests"
     )
-    private final OpStatsLogger readEntryOpLogger;
+    protected final OpStatsLogger readEntryOpLogger;
     @StatsDoc(
         name = BookKeeperClientStats.CHANNEL_TIMEOUT_READ,
         help = "timeout stats of read entries requests"
     )
-    private final OpStatsLogger readTimeoutOpLogger;
+    protected final OpStatsLogger readTimeoutOpLogger;
     @StatsDoc(
         name = BookKeeperClientStats.CHANNEL_ADD_OP,
         help = "channel stats of add entries requests"
     )
-    private final OpStatsLogger addEntryOpLogger;
+    protected final OpStatsLogger addEntryOpLogger;
     @StatsDoc(
         name = BookKeeperClientStats.CHANNEL_WRITE_LAC_OP,
         help = "channel stats of write_lac requests"
     )
-    private final OpStatsLogger writeLacOpLogger;
+    protected final OpStatsLogger writeLacOpLogger;
     @StatsDoc(
         name = BookKeeperClientStats.CHANNEL_FORCE_OP,
         help = "channel stats of force requests"
     )
-    private final OpStatsLogger forceLedgerOpLogger;
+    protected final OpStatsLogger forceLedgerOpLogger;
     @StatsDoc(
         name = BookKeeperClientStats.CHANNEL_READ_LAC_OP,
         help = "channel stats of read_lac requests"
     )
-    private final OpStatsLogger readLacOpLogger;
+    protected final OpStatsLogger readLacOpLogger;
     @StatsDoc(
         name = BookKeeperClientStats.CHANNEL_TIMEOUT_ADD,
         help = "timeout stats of add entries requests"
     )
-    private final OpStatsLogger addTimeoutOpLogger;
+    protected final OpStatsLogger addTimeoutOpLogger;
     @StatsDoc(
         name = BookKeeperClientStats.CHANNEL_TIMEOUT_WRITE_LAC,
         help = "timeout stats of write_lac requests"
     )
-    private final OpStatsLogger writeLacTimeoutOpLogger;
+    protected final OpStatsLogger writeLacTimeoutOpLogger;
     @StatsDoc(
         name = BookKeeperClientStats.CHANNEL_TIMEOUT_FORCE,
         help = "timeout stats of force requests"
     )
-    private final OpStatsLogger forceLedgerTimeoutOpLogger;
+    protected final OpStatsLogger forceLedgerTimeoutOpLogger;
     @StatsDoc(
         name = BookKeeperClientStats.CHANNEL_TIMEOUT_READ_LAC,
         help = "timeout stats of read_lac requests"
     )
-    private final OpStatsLogger readLacTimeoutOpLogger;
+    protected final OpStatsLogger readLacTimeoutOpLogger;
     @StatsDoc(
         name = BookKeeperClientStats.GET_BOOKIE_INFO_OP,
         help = "channel stats of get_bookie_info requests"
     )
-    private final OpStatsLogger getBookieInfoOpLogger;
+    protected final OpStatsLogger getBookieInfoOpLogger;
     @StatsDoc(
         name = BookKeeperClientStats.TIMEOUT_GET_BOOKIE_INFO,
         help = "timeout stats of get_bookie_info requests"
     )
-    private final OpStatsLogger getBookieInfoTimeoutOpLogger;
+    protected final OpStatsLogger getBookieInfoTimeoutOpLogger;
     @StatsDoc(
         name = BookKeeperClientStats.CHANNEL_START_TLS_OP,
         help = "channel stats of start_tls requests"
     )
-    private final OpStatsLogger startTLSOpLogger;
+    protected final OpStatsLogger startTLSOpLogger;
     @StatsDoc(
         name = BookKeeperClientStats.CHANNEL_TIMEOUT_START_TLS_OP,
         help = "timeout stats of start_tls requests"
     )
-    private final OpStatsLogger startTLSTimeoutOpLogger;
+    protected final OpStatsLogger startTLSTimeoutOpLogger;
     @StatsDoc(
         name = BookKeeperClientStats.CLIENT_CONNECT_TIMER,
         help = "channel stats of connect requests"
     )
     private final OpStatsLogger connectTimer;
-    private final OpStatsLogger getListOfEntriesOfLedgerCompletionOpLogger;
-    private final OpStatsLogger getListOfEntriesOfLedgerCompletionTimeoutOpLogger;
+    protected final OpStatsLogger getListOfEntriesOfLedgerCompletionOpLogger;
+    protected final OpStatsLogger getListOfEntriesOfLedgerCompletionTimeoutOpLogger;
     @StatsDoc(
         name = BookKeeperClientStats.NETTY_EXCEPTION_CNT,
         help = "the number of exceptions received from this channel"
@@ -283,12 +283,12 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
         name = BookKeeperClientStats.ADD_OP_OUTSTANDING,
         help = "the number of outstanding add_entry requests"
     )
-    private final Counter addEntryOutstanding;
+    protected final Counter addEntryOutstanding;
     @StatsDoc(
         name = BookKeeperClientStats.READ_OP_OUTSTANDING,
         help = "the number of outstanding add_entry requests"
     )
-    private final Counter readEntryOutstanding;
+    protected final Counter readEntryOutstanding;
     /* collect stats on all Ops that flows through netty pipeline */
     @StatsDoc(
         name = BookKeeperClientStats.NETTY_OPS,
@@ -317,7 +317,7 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
     private final Counter failedTlsHandshakeCounter;
 
     private final boolean useV2WireProtocol;
-    private final boolean preserveMdcForTaskExecution;
+    protected final boolean preserveMdcForTaskExecution;
 
     /**
      * The following member variables do not need to be concurrent, or volatile
@@ -343,7 +343,7 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
 
     volatile ConnectionState state;
     final ReentrantReadWriteLock closeLock = new ReentrantReadWriteLock();
-    private final ClientConfiguration conf;
+    protected final ClientConfiguration conf;
 
     private final PerChannelBookieClientPool pcbcPool;
     private final ClientAuthProvider.Factory authProviderFactory;
@@ -698,7 +698,7 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
         // writeLac is mostly like addEntry hence uses addEntryTimeout
         completionObjects.put(completionKey,
                               new WriteLacCompletion(completionKey, cb,
-                                                     ctx, ledgerId));
+                                                     ctx, ledgerId, this));
 
         // Build the request
         BKPacketHeader.Builder headerBuilder = BKPacketHeader.newBuilder()
@@ -737,7 +737,7 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
         // force is mostly like addEntry hence uses addEntryTimeout
         completionObjects.put(completionKey,
                               new ForceLedgerCompletion(completionKey, cb,
-                                                     ctx, ledgerId));
+                                                     ctx, ledgerId, this));
 
         // Build the request
         BKPacketHeader.Builder headerBuilder = BKPacketHeader.newBuilder()
@@ -787,7 +787,7 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
                 cb.writeComplete(BKException.Code.IllegalOpException, ledgerId, entryId, bookieId, ctx);
                 return;
             }
-            completionKey = acquireV2Key(ledgerId, entryId, OperationType.ADD_ENTRY);
+            completionKey = EntryCompletionKey.acquireV2Key(ledgerId, entryId, OperationType.ADD_ENTRY);
 
             if (toSend instanceof ByteBuf) {
                 ByteBuf byteBuf = ((ByteBuf) toSend).retainedDuplicate();
@@ -839,8 +839,8 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
         }
 
         putCompletionKeyValue(completionKey,
-                              acquireAddCompletion(completionKey,
-                                                   cb, ctx, ledgerId, entryId));
+                              AddCompletion.acquireAddCompletion(completionKey,
+                                                   cb, ctx, ledgerId, entryId, this));
         // addEntry times out on backpressure
         writeAndFlush(channel, completionKey, request, allowFastFail, cleanupActionFailedBeforeWrite,
                 cleanupActionAfterWrite);
@@ -852,7 +852,7 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
         if (useV2WireProtocol) {
             request = BookieProtocol.ReadRequest.create(BookieProtocol.CURRENT_PROTOCOL_VERSION,
                                                      ledgerId, 0, (short) 0, null);
-            completionKey = acquireV2Key(ledgerId, 0, OperationType.READ_LAC);
+            completionKey = EntryCompletionKey.acquireV2Key(ledgerId, 0, OperationType.READ_LAC);
         } else {
             final long txnId = getTxnId();
             completionKey = new TxnCompletionKey(txnId, OperationType.READ_LAC);
@@ -871,14 +871,15 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
         }
         putCompletionKeyValue(completionKey,
                               new ReadLacCompletion(completionKey, cb,
-                                                    ctx, ledgerId));
+                                                    ctx, ledgerId, this));
         writeAndFlush(channel, completionKey, request);
     }
 
     public void getListOfEntriesOfLedger(final long ledgerId, GetListOfEntriesOfLedgerCallback cb) {
         final long txnId = getTxnId();
         final CompletionKey completionKey = new TxnCompletionKey(txnId, OperationType.GET_LIST_OF_ENTRIES_OF_LEDGER);
-        completionObjects.put(completionKey, new GetListOfEntriesOfLedgerCompletion(completionKey, cb, ledgerId));
+        completionObjects.put(completionKey, new GetListOfEntriesOfLedgerCompletion(
+                completionKey, cb, ledgerId, this));
 
         // Build the request.
         BKPacketHeader.Builder headerBuilder = BKPacketHeader.newBuilder().setVersion(ProtocolVersion.VERSION_THREE)
@@ -936,7 +937,7 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
         if (useV2WireProtocol) {
             request = BookieProtocol.ReadRequest.create(BookieProtocol.CURRENT_PROTOCOL_VERSION,
                     ledgerId, entryId, (short) flags, masterKey);
-            completionKey = acquireV2Key(ledgerId, entryId, OperationType.READ_ENTRY);
+            completionKey = EntryCompletionKey.acquireV2Key(ledgerId, entryId, OperationType.READ_ENTRY);
         } else {
             final long txnId = getTxnId();
             completionKey = new TxnCompletionKey(txnId, OperationType.READ_ENTRY);
@@ -995,7 +996,7 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
                     .build();
         }
 
-        ReadCompletion readCompletion = new ReadCompletion(completionKey, cb, ctx, ledgerId, entryId);
+        ReadCompletion readCompletion = new ReadCompletion(completionKey, cb, ctx, ledgerId, entryId, this);
         putCompletionKeyValue(completionKey, readCompletion);
 
         writeAndFlush(channel, completionKey, request, allowFastFail, null, null);
@@ -1038,7 +1039,7 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
             throw new UnsupportedOperationException("Unsupported batch read entry operation for v3 protocol.");
         }
         BatchedReadCompletion readCompletion = new BatchedReadCompletion(
-                completionKey, cb, ctx, ledgerId, startEntryId);
+                completionKey, cb, ctx, ledgerId, startEntryId, this);
         putCompletionKeyValue(completionKey, readCompletion);
 
         writeAndFlush(channel, completionKey, request, allowFastFail, null, null);
@@ -1049,7 +1050,7 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
         final CompletionKey completionKey = new TxnCompletionKey(txnId, OperationType.GET_BOOKIE_INFO);
         completionObjects.put(completionKey,
                               new GetBookieInfoCompletion(
-                                      completionKey, cb, ctx));
+                                      completionKey, cb, ctx, this));
 
         // Build the request and calculate the total size to be included in the packet.
         BKPacketHeader.Builder headerBuilder = BKPacketHeader.newBuilder()
@@ -1413,7 +1414,7 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
         if (OperationType.BATCH_READ_ENTRY == operationType) {
             key = new TxnCompletionKey(((BookieProtocol.BatchedReadResponse) response).getRequestId(), operationType);
         } else {
-            key = acquireV2Key(response.ledgerId, response.entryId, operationType);
+            key = EntryCompletionKey.acquireV2Key(response.ledgerId, response.entryId, operationType);
         }
         CompletionValue completionValue = getCompletionValue(key);
         key.release();
@@ -1656,811 +1657,9 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
             });
     }
 
-    /**
-     * Boiler-plate wrapper classes follow.
-     *
-     */
-
-    // visible for testing
-    abstract class CompletionValue {
-        private final OpStatsLogger opLogger;
-        private final OpStatsLogger timeoutOpLogger;
-        private final String operationName;
-        private final Map<String, String> mdcContextMap;
-        protected Object ctx;
-        protected long ledgerId;
-        protected long entryId;
-        protected long startTime;
-
-        public CompletionValue(String operationName,
-                               Object ctx,
-                               long ledgerId, long entryId,
-                               OpStatsLogger opLogger,
-                               OpStatsLogger timeoutOpLogger) {
-            this.operationName = operationName;
-            this.ctx = ctx;
-            this.ledgerId = ledgerId;
-            this.entryId = entryId;
-            this.startTime = MathUtils.nowInNano();
-            this.opLogger = opLogger;
-            this.timeoutOpLogger = timeoutOpLogger;
-            this.mdcContextMap = preserveMdcForTaskExecution ? MDC.getCopyOfContextMap() : null;
-        }
-
-        private long latency() {
-            return MathUtils.elapsedNanos(startTime);
-        }
-
-        void logOpResult(int rc) {
-            if (rc != BKException.Code.OK) {
-                opLogger.registerFailedEvent(latency(), TimeUnit.NANOSECONDS);
-            } else {
-                opLogger.registerSuccessfulEvent(latency(), TimeUnit.NANOSECONDS);
-            }
-
-            if (rc != BKException.Code.OK
-                && !expectedBkOperationErrors.contains(rc)) {
-                recordError();
-            }
-        }
-
-        boolean maybeTimeout() {
-            if (MathUtils.elapsedNanos(startTime) >= readEntryTimeoutNanos) {
-                timeout();
-                return true;
-            } else {
-                return false;
-            }
-        }
-
-        void timeout() {
-            errorOut(BKException.Code.TimeoutException);
-            timeoutOpLogger.registerSuccessfulEvent(latency(),
-                                                    TimeUnit.NANOSECONDS);
-        }
-
-        protected void logResponse(StatusCode status, Object... extraInfo) {
-            if (LOG.isDebugEnabled()) {
-                LOG.debug("Got {} response from bookie:{} rc:{}, {}", operationName, bookieId, status,
-                        Joiner.on(":").join(extraInfo));
-            }
-        }
-
-        protected int convertStatus(StatusCode status, int defaultStatus) {
-            // convert to BKException code
-            int rcToRet = statusCodeToExceptionCode(status);
-            if (rcToRet == BKException.Code.UNINITIALIZED) {
-                LOG.error("{} for failed on bookie {} code {}",
-                          operationName, bookieId, status);
-                return defaultStatus;
-            } else {
-                return rcToRet;
-            }
-        }
-
-        public void restoreMdcContext() {
-            MdcUtils.restoreContext(mdcContextMap);
-        }
-
-        public abstract void errorOut();
-        public abstract void errorOut(int rc);
-        public void setOutstanding() {
-            // no-op
-        }
-
-        protected void errorOutAndRunCallback(final Runnable callback) {
-            executor.executeOrdered(ledgerId, () -> {
-                String bAddress = "null";
-                Channel c = channel;
-                if (c != null && c.remoteAddress() != null) {
-                    bAddress = c.remoteAddress().toString();
-                }
-                if (LOG.isDebugEnabled()) {
-                    LOG.debug("Could not write {} request to bookie {} for ledger {}, entry {}",
-                            operationName, bAddress,
-                            ledgerId, entryId);
-                }
-                callback.run();
-            });
-        }
-
-        public void handleV2Response(
-                long ledgerId, long entryId, StatusCode status,
-                BookieProtocol.Response response) {
-            LOG.warn("Unhandled V2 response {}", response);
-        }
-
-        public abstract void handleV3Response(
-                BookkeeperProtocol.Response response);
-    }
-
-    // visible for testing
-    class WriteLacCompletion extends CompletionValue {
-        final WriteLacCallback cb;
-
-        public WriteLacCompletion(final CompletionKey key,
-                                  final WriteLacCallback originalCallback,
-                                  final Object originalCtx,
-                                  final long ledgerId) {
-            super("WriteLAC",
-                  originalCtx, ledgerId, BookieProtocol.LAST_ADD_CONFIRMED,
-                  writeLacOpLogger, writeLacTimeoutOpLogger);
-            this.cb = new WriteLacCallback() {
-                    @Override
-                    public void writeLacComplete(int rc, long ledgerId,
-                                                 BookieId addr,
-                                                 Object ctx) {
-                        logOpResult(rc);
-                        originalCallback.writeLacComplete(rc, ledgerId,
-                                                          addr, originalCtx);
-                        key.release();
-                    }
-                };
-        }
-
-        @Override
-        public void errorOut() {
-            errorOut(BKException.Code.BookieHandleNotAvailableException);
-        }
-
-        @Override
-        public void errorOut(final int rc) {
-            errorOutAndRunCallback(
-                    () -> cb.writeLacComplete(rc, ledgerId, bookieId, ctx));
-        }
-
-        @Override
-        public void handleV3Response(BookkeeperProtocol.Response response) {
-            WriteLacResponse writeLacResponse = response.getWriteLacResponse();
-            StatusCode status = response.getStatus() == StatusCode.EOK
-                ? writeLacResponse.getStatus() : response.getStatus();
-            long ledgerId = writeLacResponse.getLedgerId();
-
-            if (LOG.isDebugEnabled()) {
-                logResponse(status, "ledger", ledgerId);
-            }
-            int rc = convertStatus(status, BKException.Code.WriteException);
-            cb.writeLacComplete(rc, ledgerId, bookieId, ctx);
-        }
-    }
-
-    class ForceLedgerCompletion extends CompletionValue {
-        final ForceLedgerCallback cb;
-
-        public ForceLedgerCompletion(final CompletionKey key,
-                                  final ForceLedgerCallback originalCallback,
-                                  final Object originalCtx,
-                                  final long ledgerId) {
-            super("ForceLedger",
-                  originalCtx, ledgerId, BookieProtocol.LAST_ADD_CONFIRMED,
-                  forceLedgerOpLogger, forceLedgerTimeoutOpLogger);
-            this.cb = new ForceLedgerCallback() {
-                    @Override
-                    public void forceLedgerComplete(int rc, long ledgerId,
-                                                 BookieId addr,
-                                                 Object ctx) {
-                        logOpResult(rc);
-                        originalCallback.forceLedgerComplete(rc, ledgerId,
-                                                          addr, originalCtx);
-                        key.release();
-                    }
-                };
-        }
-
-        @Override
-        public void errorOut() {
-            errorOut(BKException.Code.BookieHandleNotAvailableException);
-        }
-
-        @Override
-        public void errorOut(final int rc) {
-            errorOutAndRunCallback(
-                    () -> cb.forceLedgerComplete(rc, ledgerId, bookieId, ctx));
-        }
-
-        @Override
-        public void handleV3Response(BookkeeperProtocol.Response response) {
-            ForceLedgerResponse forceLedgerResponse = response.getForceLedgerResponse();
-            StatusCode status = response.getStatus() == StatusCode.EOK
-                ? forceLedgerResponse.getStatus() : response.getStatus();
-            long ledgerId = forceLedgerResponse.getLedgerId();
-
-            if (LOG.isDebugEnabled()) {
-                logResponse(status, "ledger", ledgerId);
-            }
-            int rc = convertStatus(status, BKException.Code.WriteException);
-            cb.forceLedgerComplete(rc, ledgerId, bookieId, ctx);
-        }
-    }
-
-    // visible for testing
-    class ReadLacCompletion extends CompletionValue {
-        final ReadLacCallback cb;
-
-        public ReadLacCompletion(final CompletionKey key,
-                                 ReadLacCallback originalCallback,
-                                 final Object ctx, final long ledgerId) {
-            super("ReadLAC", ctx, ledgerId, BookieProtocol.LAST_ADD_CONFIRMED,
-                  readLacOpLogger, readLacTimeoutOpLogger);
-            this.cb = new ReadLacCallback() {
-                    @Override
-                    public void readLacComplete(int rc, long ledgerId,
-                                                ByteBuf lacBuffer,
-                                                ByteBuf lastEntryBuffer,
-                                                Object ctx) {
-                        logOpResult(rc);
-                        originalCallback.readLacComplete(
-                                rc, ledgerId, lacBuffer, lastEntryBuffer, ctx);
-                        key.release();
-                    }
-                };
-        }
-
-        @Override
-        public void errorOut() {
-            errorOut(BKException.Code.BookieHandleNotAvailableException);
-        }
-
-        @Override
-        public void errorOut(final int rc) {
-            errorOutAndRunCallback(
-                    () -> cb.readLacComplete(rc, ledgerId, null, null, ctx));
-        }
-
-        @Override
-        public void handleV3Response(BookkeeperProtocol.Response response) {
-            ReadLacResponse readLacResponse = response.getReadLacResponse();
-            ByteBuf lacBuffer = Unpooled.EMPTY_BUFFER;
-            ByteBuf lastEntryBuffer = Unpooled.EMPTY_BUFFER;
-            StatusCode status = response.getStatus() == StatusCode.EOK
-                ? readLacResponse.getStatus() : response.getStatus();
-
-            if (readLacResponse.hasLacBody()) {
-                lacBuffer = Unpooled.wrappedBuffer(readLacResponse.getLacBody().asReadOnlyByteBuffer());
-            }
-
-            if (readLacResponse.hasLastEntryBody()) {
-                lastEntryBuffer = Unpooled.wrappedBuffer(readLacResponse.getLastEntryBody().asReadOnlyByteBuffer());
-            }
-
-            if (LOG.isDebugEnabled()) {
-                logResponse(status, "ledgerId", ledgerId);
-            }
-
-            int rc = convertStatus(status, BKException.Code.ReadException);
-            cb.readLacComplete(rc, ledgerId, lacBuffer.slice(),
-                               lastEntryBuffer.slice(), ctx);
-        }
-    }
-
-    // visible for testing
-    class ReadCompletion extends CompletionValue {
-        final ReadEntryCallback cb;
-
-        public ReadCompletion(final CompletionKey key,
-                              final ReadEntryCallback originalCallback,
-                              final Object originalCtx,
-                              long ledgerId, final long entryId) {
-            super("Read", originalCtx, ledgerId, entryId,
-                  readEntryOpLogger, readTimeoutOpLogger);
-
-            this.cb = new ReadEntryCallback() {
-                    @Override
-                    public void readEntryComplete(int rc, long ledgerId,
-                                                  long entryId, ByteBuf buffer,
-                                                  Object ctx) {
-                        logOpResult(rc);
-                        originalCallback.readEntryComplete(rc,
-                                                           ledgerId, entryId,
-                                                           buffer, originalCtx);
-                        key.release();
-                    }
-                };
-        }
-
-        @Override
-        public void errorOut() {
-            errorOut(BKException.Code.BookieHandleNotAvailableException);
-        }
-
-        @Override
-        public void errorOut(final int rc) {
-            errorOutAndRunCallback(
-                    () -> cb.readEntryComplete(rc, ledgerId,
-                                               entryId, null, ctx));
-        }
-
-        @Override
-        public void setOutstanding() {
-            readEntryOutstanding.inc();
-        }
-
-        @Override
-        public void handleV2Response(long ledgerId, long entryId,
-                                     StatusCode status,
-                                     BookieProtocol.Response response) {
-            readEntryOutstanding.dec();
-            if (!(response instanceof BookieProtocol.ReadResponse)) {
-                return;
-            }
-            BookieProtocol.ReadResponse readResponse = (BookieProtocol.ReadResponse) response;
-            handleReadResponse(ledgerId, entryId, status, readResponse.getData(),
-                               INVALID_ENTRY_ID, -1L);
-        }
-
-        @Override
-        public void handleV3Response(BookkeeperProtocol.Response response) {
-            readEntryOutstanding.dec();
-            ReadResponse readResponse = response.getReadResponse();
-            StatusCode status = response.getStatus() == StatusCode.EOK
-                ? readResponse.getStatus() : response.getStatus();
-            ByteBuf buffer = Unpooled.EMPTY_BUFFER;
-            if (readResponse.hasBody()) {
-                buffer = Unpooled.wrappedBuffer(readResponse.getBody().asReadOnlyByteBuffer());
-            }
-            long maxLAC = INVALID_ENTRY_ID;
-            if (readResponse.hasMaxLAC()) {
-                maxLAC = readResponse.getMaxLAC();
-            }
-            long lacUpdateTimestamp = -1L;
-            if (readResponse.hasLacUpdateTimestamp()) {
-                lacUpdateTimestamp = readResponse.getLacUpdateTimestamp();
-            }
-            handleReadResponse(readResponse.getLedgerId(),
-                               readResponse.getEntryId(),
-                               status, buffer, maxLAC, lacUpdateTimestamp);
-            ReferenceCountUtil.release(
-                    buffer); // meaningless using unpooled, but client may expect to hold the last reference
-        }
-
-        private void handleReadResponse(long ledgerId,
-                                        long entryId,
-                                        StatusCode status,
-                                        ByteBuf buffer,
-                                        long maxLAC, // max known lac piggy-back from bookies
-                                        long lacUpdateTimestamp) { // the timestamp when the lac is updated.
-            int readableBytes = buffer.readableBytes();
-            if (LOG.isDebugEnabled()) {
-                logResponse(status, "ledger", ledgerId, "entry", entryId, "entryLength", readableBytes);
-            }
-
-            int rc = convertStatus(status, BKException.Code.ReadException);
-
-            if (maxLAC > INVALID_ENTRY_ID && (ctx instanceof ReadEntryCallbackCtx)) {
-                ((ReadEntryCallbackCtx) ctx).setLastAddConfirmed(maxLAC);
-            }
-            if (lacUpdateTimestamp > -1L && (ctx instanceof ReadLastConfirmedAndEntryContext)) {
-                ((ReadLastConfirmedAndEntryContext) ctx).setLacUpdateTimestamp(lacUpdateTimestamp);
-            }
-            cb.readEntryComplete(rc, ledgerId, entryId, buffer.slice(), ctx);
-        }
-    }
-
-    class BatchedReadCompletion extends CompletionValue {
-
-        final BatchedReadEntryCallback cb;
-
-        public BatchedReadCompletion(final CompletionKey key,
-                                     final BatchedReadEntryCallback originalCallback,
-                                     final Object originalCtx,
-                                     long ledgerId, final long entryId) {
-            super("BatchedRead", originalCtx, ledgerId, entryId,
-                    readEntryOpLogger, readTimeoutOpLogger);
-            this.cb = new BatchedReadEntryCallback() {
-
-                @Override
-                public void readEntriesComplete(int rc,
-                                                long ledgerId,
-                                                long startEntryId,
-                                                ByteBufList bufList,
-                                                Object ctx) {
-                    logOpResult(rc);
-                    originalCallback.readEntriesComplete(rc,
-                            ledgerId, entryId,
-                            bufList, originalCtx);
-                    key.release();
-                }
-            };
-        }
-
-        @Override
-        public void errorOut() {
-            errorOut(BKException.Code.BookieHandleNotAvailableException);
-        }
-
-        @Override
-        public void errorOut(final int rc) {
-            errorOutAndRunCallback(
-                    () -> cb.readEntriesComplete(rc, ledgerId,
-                            entryId, null, ctx));
-        }
-
-        @Override
-        public void handleV2Response(long ledgerId,
-                                     long entryId,
-                                     StatusCode status,
-                                     BookieProtocol.Response response) {
-
-            readEntryOutstanding.dec();
-            if (!(response instanceof BookieProtocol.BatchedReadResponse)) {
-                return;
-            }
-            BookieProtocol.BatchedReadResponse readResponse = (BookieProtocol.BatchedReadResponse) response;
-            handleBatchedReadResponse(ledgerId, entryId, status, readResponse.getData(),
-                    INVALID_ENTRY_ID, -1L);
-        }
-
-        @Override
-        public void handleV3Response(Response response) {
-            // V3 protocol haven't supported batched read yet.
-        }
-
-        private void handleBatchedReadResponse(long ledgerId,
-                                        long entryId,
-                                        StatusCode status,
-                                        ByteBufList buffers,
-                                        long maxLAC, // max known lac piggy-back from bookies
-                                        long lacUpdateTimestamp) { // the timestamp when the lac is updated.
-            int rc = convertStatus(status, BKException.Code.ReadException);
-
-            if (maxLAC > INVALID_ENTRY_ID && (ctx instanceof ReadEntryCallbackCtx)) {
-                ((ReadEntryCallbackCtx) ctx).setLastAddConfirmed(maxLAC);
-            }
-            if (lacUpdateTimestamp > -1L && (ctx instanceof ReadLastConfirmedAndEntryContext)) {
-                ((ReadLastConfirmedAndEntryContext) ctx).setLacUpdateTimestamp(lacUpdateTimestamp);
-            }
-            cb.readEntriesComplete(rc, ledgerId, entryId, buffers, ctx);
-        }
-    }
-
-    class StartTLSCompletion extends CompletionValue {
-        final StartTLSCallback cb;
-
-        public StartTLSCompletion(final CompletionKey key) {
-            super("StartTLS", null, -1, -1,
-                  startTLSOpLogger, startTLSTimeoutOpLogger);
-            this.cb = new StartTLSCallback() {
-                @Override
-                public void startTLSComplete(int rc, Object ctx) {
-                    logOpResult(rc);
-                    key.release();
-                }
-            };
-        }
-
-        @Override
-        public void errorOut() {
-            errorOut(BKException.Code.BookieHandleNotAvailableException);
-        }
-
-        @Override
-        public void errorOut(final int rc) {
-            failTLS(rc);
-        }
-
-        @Override
-        public void handleV3Response(BookkeeperProtocol.Response response) {
-            StatusCode status = response.getStatus();
-
-            if (LOG.isDebugEnabled()) {
-                logResponse(status);
-            }
-
-            int rc = convertStatus(status, BKException.Code.SecurityException);
-
-            // Cancel START_TLS request timeout
-            cb.startTLSComplete(rc, null);
-
-            if (state != ConnectionState.START_TLS) {
-                LOG.error("Connection state changed before TLS response received");
-                failTLS(BKException.Code.BookieHandleNotAvailableException);
-            } else if (status != StatusCode.EOK) {
-                LOG.error("Client received error {} during TLS negotiation", status);
-                failTLS(BKException.Code.SecurityException);
-            } else {
-                initTLSHandshake();
-            }
-        }
-
-    }
-
-    // visible for testing
-    class GetBookieInfoCompletion extends CompletionValue {
-        final GetBookieInfoCallback cb;
-
-        public GetBookieInfoCompletion(final CompletionKey key,
-                                       final GetBookieInfoCallback origCallback,
-                                       final Object origCtx) {
-            super("GetBookieInfo", origCtx, 0L, 0L,
-                  getBookieInfoOpLogger, getBookieInfoTimeoutOpLogger);
-            this.cb = new GetBookieInfoCallback() {
-                @Override
-                public void getBookieInfoComplete(int rc, BookieInfo bInfo,
-                                                  Object ctx) {
-                    logOpResult(rc);
-                    origCallback.getBookieInfoComplete(rc, bInfo, origCtx);
-                    key.release();
-                }
-            };
-        }
-
-        @Override
-        boolean maybeTimeout() {
-            if (MathUtils.elapsedNanos(startTime) >= getBookieInfoTimeoutNanos) {
-                timeout();
-                return true;
-            } else {
-                return false;
-            }
-        }
-
-        @Override
-        public void errorOut() {
-            errorOut(BKException.Code.BookieHandleNotAvailableException);
-        }
-
-        @Override
-        public void errorOut(final int rc) {
-            errorOutAndRunCallback(
-                    () -> cb.getBookieInfoComplete(rc, new BookieInfo(), ctx));
-        }
-
-        @Override
-        public void handleV3Response(BookkeeperProtocol.Response response) {
-            GetBookieInfoResponse getBookieInfoResponse = response.getGetBookieInfoResponse();
-            StatusCode status = response.getStatus() == StatusCode.EOK
-                ? getBookieInfoResponse.getStatus() : response.getStatus();
-
-            long freeDiskSpace = getBookieInfoResponse.getFreeDiskSpace();
-            long totalDiskSpace = getBookieInfoResponse.getTotalDiskCapacity();
-
-            if (LOG.isDebugEnabled()) {
-                logResponse(status, "freeDisk", freeDiskSpace, "totalDisk", totalDiskSpace);
-            }
-
-            int rc = convertStatus(status, BKException.Code.ReadException);
-            cb.getBookieInfoComplete(rc,
-                                     new BookieInfo(totalDiskSpace,
-                                                    freeDiskSpace), ctx);
-        }
-    }
-
-    class GetListOfEntriesOfLedgerCompletion extends CompletionValue {
-        final GetListOfEntriesOfLedgerCallback cb;
-
-        public GetListOfEntriesOfLedgerCompletion(final CompletionKey key,
-                final GetListOfEntriesOfLedgerCallback origCallback, final long ledgerId) {
-            super("GetListOfEntriesOfLedger", null, ledgerId, 0L, getListOfEntriesOfLedgerCompletionOpLogger,
-                    getListOfEntriesOfLedgerCompletionTimeoutOpLogger);
-            this.cb = new GetListOfEntriesOfLedgerCallback() {
-                @Override
-                public void getListOfEntriesOfLedgerComplete(int rc, long ledgerId,
-                        AvailabilityOfEntriesOfLedger availabilityOfEntriesOfLedger) {
-                    logOpResult(rc);
-                    origCallback.getListOfEntriesOfLedgerComplete(rc, ledgerId, availabilityOfEntriesOfLedger);
-                    key.release();
-                }
-            };
-        }
-
-        @Override
-        public void errorOut() {
-            errorOut(BKException.Code.BookieHandleNotAvailableException);
-        }
-
-        @Override
-        public void errorOut(final int rc) {
-            errorOutAndRunCallback(() -> cb.getListOfEntriesOfLedgerComplete(rc, ledgerId, null));
-        }
-
-        @Override
-        public void handleV3Response(BookkeeperProtocol.Response response) {
-            GetListOfEntriesOfLedgerResponse getListOfEntriesOfLedgerResponse = response
-                    .getGetListOfEntriesOfLedgerResponse();
-            ByteBuf availabilityOfEntriesOfLedgerBuffer = Unpooled.EMPTY_BUFFER;
-            StatusCode status = response.getStatus() == StatusCode.EOK ? getListOfEntriesOfLedgerResponse.getStatus()
-                    : response.getStatus();
-
-            if (getListOfEntriesOfLedgerResponse.hasAvailabilityOfEntriesOfLedger()) {
-                availabilityOfEntriesOfLedgerBuffer = Unpooled.wrappedBuffer(
-                        getListOfEntriesOfLedgerResponse.getAvailabilityOfEntriesOfLedger().asReadOnlyByteBuffer());
-            }
-
-            if (LOG.isDebugEnabled()) {
-                logResponse(status, "ledgerId", ledgerId);
-            }
-
-            int rc = convertStatus(status, BKException.Code.ReadException);
-            AvailabilityOfEntriesOfLedger availabilityOfEntriesOfLedger = null;
-            if (rc == BKException.Code.OK) {
-                availabilityOfEntriesOfLedger = new AvailabilityOfEntriesOfLedger(
-                        availabilityOfEntriesOfLedgerBuffer.slice());
-            }
-            cb.getListOfEntriesOfLedgerComplete(rc, ledgerId, availabilityOfEntriesOfLedger);
-        }
-    }
-
-    private final Recycler<AddCompletion> addCompletionRecycler = new Recycler<AddCompletion>() {
-            @Override
-            protected AddCompletion newObject(Recycler.Handle<AddCompletion> handle) {
-                return new AddCompletion(handle);
-            }
-        };
-
-    AddCompletion acquireAddCompletion(final CompletionKey key,
-                                       final WriteCallback originalCallback,
-                                       final Object originalCtx,
-                                       final long ledgerId, final long entryId) {
-        AddCompletion completion = addCompletionRecycler.get();
-        completion.reset(key, originalCallback, originalCtx, ledgerId, entryId);
-        return completion;
-    }
-
-    // visible for testing
-    class AddCompletion extends CompletionValue implements WriteCallback {
-        final Recycler.Handle<AddCompletion> handle;
-
-        CompletionKey key = null;
-        WriteCallback originalCallback = null;
-
-        AddCompletion(Recycler.Handle<AddCompletion> handle) {
-            super("Add", null, -1, -1, addEntryOpLogger, addTimeoutOpLogger);
-            this.handle = handle;
-        }
-
-        void reset(final CompletionKey key,
-                   final WriteCallback originalCallback,
-                   final Object originalCtx,
-                   final long ledgerId, final long entryId) {
-            this.key = key;
-            this.originalCallback = originalCallback;
-            this.ctx = originalCtx;
-            this.ledgerId = ledgerId;
-            this.entryId = entryId;
-            this.startTime = MathUtils.nowInNano();
-        }
-
-        @Override
-        public void writeComplete(int rc, long ledgerId, long entryId,
-                                  BookieId addr,
-                                  Object ctx) {
-            logOpResult(rc);
-            originalCallback.writeComplete(rc, ledgerId, entryId, addr, ctx);
-            key.release();
-            handle.recycle(this);
-        }
-
-        @Override
-        boolean maybeTimeout() {
-            if (MathUtils.elapsedNanos(startTime) >= addEntryTimeoutNanos) {
-                timeout();
-                return true;
-            } else {
-                return false;
-            }
-        }
-
-        @Override
-        public void errorOut() {
-            errorOut(BKException.Code.BookieHandleNotAvailableException);
-        }
-
-        @Override
-        public void errorOut(final int rc) {
-            errorOutAndRunCallback(
-                    () -> writeComplete(rc, ledgerId, entryId, bookieId, ctx));
-        }
-
-        @Override
-        public void setOutstanding() {
-            addEntryOutstanding.inc();
-        }
-
-        @Override
-        public void handleV2Response(
-                long ledgerId, long entryId, StatusCode status,
-                BookieProtocol.Response response) {
-            addEntryOutstanding.dec();
-            handleResponse(ledgerId, entryId, status);
-        }
-
-        @Override
-        public void handleV3Response(
-                BookkeeperProtocol.Response response) {
-            addEntryOutstanding.dec();
-            AddResponse addResponse = response.getAddResponse();
-            StatusCode status = response.getStatus() == StatusCode.EOK
-                ? addResponse.getStatus() : response.getStatus();
-            handleResponse(addResponse.getLedgerId(), addResponse.getEntryId(),
-                           status);
-        }
-
-        private void handleResponse(long ledgerId, long entryId,
-                                    StatusCode status) {
-            if (LOG.isDebugEnabled()) {
-                logResponse(status, "ledger", ledgerId, "entry", entryId);
-            }
-
-            int rc = convertStatus(status, BKException.Code.WriteException);
-            writeComplete(rc, ledgerId, entryId, bookieId, ctx);
-        }
-    }
-
     // visible for testing
     CompletionKey newCompletionKey(long txnId, OperationType operationType) {
         return new TxnCompletionKey(txnId, operationType);
-    }
-
-    class TxnCompletionKey extends CompletionKey {
-        final long txnId;
-
-        public TxnCompletionKey(long txnId, OperationType operationType) {
-            super(operationType);
-            this.txnId = txnId;
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (!(obj instanceof TxnCompletionKey)) {
-                return false;
-            }
-            TxnCompletionKey that = (TxnCompletionKey) obj;
-            return this.txnId == that.txnId && this.operationType == that.operationType;
-        }
-
-        @Override
-        public int hashCode() {
-            return ((int) txnId);
-        }
-
-        @Override
-        public String toString() {
-            return String.format("TxnId(%d), OperationType(%s)", txnId, operationType);
-        }
-
-    }
-
-    abstract class CompletionKey {
-        OperationType operationType;
-
-        CompletionKey(OperationType operationType) {
-            this.operationType = operationType;
-        }
-
-        public void release() {}
-    }
-
-    /**
-     * Note : Helper functions follow
-     */
-
-    /**
-     * @param status
-     * @return {@link BKException.Code.UNINITIALIZED} if the statuscode is unknown.
-     */
-    private int statusCodeToExceptionCode(StatusCode status) {
-        switch (status) {
-            case EOK:
-                return BKException.Code.OK;
-            case ENOENTRY:
-                return BKException.Code.NoSuchEntryException;
-            case ENOLEDGER:
-                return BKException.Code.NoSuchLedgerExistsException;
-            case EBADVERSION:
-                return BKException.Code.ProtocolVersionException;
-            case EUA:
-                return BKException.Code.UnauthorizedAccessException;
-            case EFENCED:
-                return BKException.Code.LedgerFencedException;
-            case EREADONLY:
-                return BKException.Code.WriteOnReadOnlyBookieException;
-            case ETOOMANYREQUESTS:
-                return BKException.Code.TooManyRequestsException;
-            case EUNKNOWNLEDGERSTATE:
-                return BKException.Code.DataUnknownException;
-            default:
-                return BKException.Code.UNINITIALIZED;
-        }
     }
 
     private void putCompletionKeyValue(CompletionKey key, CompletionValue value) {
@@ -2482,64 +1681,6 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
 
     private long getTxnId() {
         return txnIdGenerator.incrementAndGet();
-    }
-
-    private final Recycler<EntryCompletionKey> v2KeyRecycler = new Recycler<EntryCompletionKey>() {
-            @Override
-            protected EntryCompletionKey newObject(
-                    Recycler.Handle<EntryCompletionKey> handle) {
-                return new EntryCompletionKey(handle);
-            }
-        };
-
-    EntryCompletionKey acquireV2Key(long ledgerId, long entryId,
-                             OperationType operationType) {
-        EntryCompletionKey key = v2KeyRecycler.get();
-        key.reset(ledgerId, entryId, operationType);
-        return key;
-    }
-
-    private class EntryCompletionKey extends CompletionKey {
-        private final Handle<EntryCompletionKey> recyclerHandle;
-        long ledgerId;
-        long entryId;
-
-        private EntryCompletionKey(Handle<EntryCompletionKey> handle) {
-            super(null);
-            this.recyclerHandle = handle;
-        }
-
-        void reset(long ledgerId, long entryId, OperationType operationType) {
-            this.ledgerId = ledgerId;
-            this.entryId = entryId;
-            this.operationType = operationType;
-        }
-
-        @Override
-        public boolean equals(Object object) {
-            if (!(object instanceof EntryCompletionKey)) {
-                return  false;
-            }
-            EntryCompletionKey that = (EntryCompletionKey) object;
-            return this.entryId == that.entryId
-                && this.ledgerId == that.ledgerId
-                && this.operationType == that.operationType;
-        }
-
-        @Override
-        public int hashCode() {
-            return Long.hashCode(ledgerId) * 31 + Long.hashCode(entryId);
-        }
-
-        @Override
-        public String toString() {
-            return String.format("%d:%d %s", ledgerId, entryId, operationType);
-        }
-
-        @Override
-        public void release() {
-            recyclerHandle.recycle(this);
-        }
     }
 
     Request.Builder withRequestContext(Request.Builder builder) {
@@ -2710,7 +1851,7 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
         final long txnId = getTxnId();
         final CompletionKey completionKey = new TxnCompletionKey(txnId, OperationType.START_TLS);
         completionObjects.put(completionKey,
-                              new StartTLSCompletion(completionKey));
+                              new StartTLSCompletion(completionKey, this));
         BookkeeperProtocol.Request.Builder h = withRequestContext(BookkeeperProtocol.Request.newBuilder());
         BKPacketHeader.Builder headerBuilder = BKPacketHeader.newBuilder()
                 .setVersion(ProtocolVersion.VERSION_THREE)
@@ -2722,7 +1863,7 @@ public class PerChannelBookieClient extends ChannelInboundHandlerAdapter {
         writeAndFlush(channel, completionKey, h.build());
     }
 
-    private void failTLS(int rc) {
+    protected void failTLS(int rc) {
         LOG.error("TLS failure on: {}, rc: {}", channel, rc);
         Queue<GenericCallback<PerChannelBookieClient>> oldPendingOps;
         synchronized (this) {
